--- src/test/java/org/lowell/apps/application/support/DatabaseCleanUp.java
@@ -1,63 +1,0 @@
-package org.lowell.apps.application.support;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import jakarta.persistence.Table;
-import jakarta.persistence.metamodel.EntityType;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-@Component
-//@ActiveProfiles("test")
-public class DatabaseCleanUp implements InitializingBean {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    private List<String> tableNames = new ArrayList<>();
-
-    @Autowired
-    private RedisTemplate<String, Object> redisTemplate;
-
-    @Override
-    public void afterPropertiesSet() {
-        List<EntityType<?>> items = entityManager.getMetamodel().getEntities().stream()
-                                                .filter(e -> e.getJavaType().getAnnotation(Entity.class) != null)
-                                                .toList();
-        for (EntityType<?> entityType : items) {
-            Table tableAnnotation = entityType.getJavaType().getAnnotation(Table.class);
-            if (tableAnnotation != null) {
-                tableNames.add(tableAnnotation.name());
-            }
-        }
-    }
-
-    @Transactional
-    public void execute() {
-        entityManager.flush();
-        entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
-
-        for (String tableName : tableNames) {
-            entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
-            entityManager.createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
-        }
-
-        entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
-    }
-
-    public void cleanRedisData() {
-        Set<String> keys = redisTemplate.keys("*");
-        if (!CollectionUtils.isEmpty(keys)) {
-            redisTemplate.delete(keys);
-        }
-    }
-}
