--- src/test/java/org/lowell/apps/infra/KafkaTestProducer.java
@@ -1,18 +1,0 @@
-package org.lowell.apps.infra;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@RequiredArgsConstructor
-@Component
-public class KafkaTestProducer {
-    private final KafkaTemplate<String, Object> kafkaTemplate;
-
-    public void sendMessage(String topic, Object message) {
-        log.info("Sending message to topic: {}", topic);
-        kafkaTemplate.send(topic, message);
-    }
-}
